Use Lock For Cache, Skip Depth Changes
