+++ Sources/CodeEditSourceEditor/LineFolding/Model/FoldRange.swift
@@ -11,14 +11,16 @@ import Foundation
 class FoldRange {
     var lineRange: ClosedRange<Int>
     var range: NSRange
+    var depth: Int
     /// Ordered array of ranges that are nested in this fold.
     var subFolds: [FoldRange]
 
     weak var parent: FoldRange?
 
-    init(lineRange: ClosedRange<Int>, range: NSRange, parent: FoldRange?, subFolds: [FoldRange]) {
+    init(lineRange: ClosedRange<Int>, range: NSRange, depth: Int, parent: FoldRange?, subFolds: [FoldRange]) {
         self.lineRange = lineRange
         self.range = range
+        self.depth = depth
         self.subFolds = subFolds
         self.parent = parent
     }
